+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Change 'crudini' package to use the newer 'python-crudini' package instead.
 - Add support for iPhone file transfers and Internet tethering.
 - Change the UEFI boot name to winesapOS.
+- Change all Flatpaks to be installed during the first-time setup.
 
 ## 3.3.0 - The Major Upgrade Update